remove verbose option. better command and error printing.
decided to get rid of the verbose option. the assignment doesn't
specify the need for it, and it was ugly trying to implemtent it. made
it so when an error is encountered, the error is passed to the head of
the command and checked before attempting to print. also added an
element to comm.h for the error message for printing purposes.